+++ src/components/widgets/LabelCollapsible.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { vRef } from '@/utils/helpers';
+import { watchedPropRef } from '@/utils/helpers';
 
 const props = withDefaults(
   defineProps<{
@@ -14,7 +14,7 @@ const props = withDefaults(
   },
 );
 
-const isExpanded = vRef(props, 'isExpanded', false);
+const isExpanded = watchedPropRef(props, 'isExpanded', false);
 
 const emit = defineEmits<{
   setExpanded: [value: boolean];
@@ -29,11 +29,11 @@ function handleClick() {
 </script>
 
 <template>
-  <div class="osim-collapsable-label" v-bind="$attrs">
+  <div class="osim-collapsible-label" v-bind="$attrs">
     <button
       :disabled="!isExpandable"
       type="button"
-      class="me-2"
+      class="me-2 osim-collapsible-toggle"
       :class="{ 'pe-none': !isExpandable }"
       @click="handleClick"
     >
@@ -55,18 +55,18 @@ function handleClick() {
 </template>
 
 <style scoped>
-.osim-collapsable-label button {
+.osim-collapsible-label button {
   border: none;
   padding: 0;
   background: transparent;
 }
 
-.osim-collapsable-label button * {
+.osim-collapsible-label button * {
   cursor: pointer;
 }
 
-.osim-collapsable-label :deep(div.osim-static-label),
-.osim-collapsable-label :deep(.osim-input) {
+.osim-collapsible-label :deep(div.osim-static-label),
+.osim-collapsible-label :deep(.osim-input) {
   padding-left: 0 !important;
   border-left: none !important;
   margin-left: 0 !important;