+++ src/composables/useFlawModel.ts
@@ -43,10 +43,15 @@ export function useFlawModel(forFlaw: ZodFlawType = blankFlaw(), onSaveSuccess: 
   const bugzillaLink = computed(() => getFlawBugzillaLink(flaw.value));
   const osimLink = computed(() => getFlawOsimLink(flaw.value.uuid));
 
+  function isValid() {
+    return ZodFlawSchema.safeParse(flaw.value).success;
+  }
+
   async function createFlaw() {
     isSaving.value = true;
     const validatedFlaw = validate();
     if (!validatedFlaw.success) {
+      isSaving.value = false;
       return;
     }
     // Remove any empty fields before request
@@ -141,6 +146,7 @@ export function useFlawModel(forFlaw: ZodFlawType = blankFlaw(), onSaveSuccess: 
   return {
     flaw,
     isSaving,
+    isValid,
     errors,
     committedFlaw,
     trackerUuids,